Using Eigen as a contiguously-allocated tensor to improve performance.

+++ source/raytracer/include/raytracer/alias.hpp
@@ -26,5 +26,5 @@
 namespace raytracer
 {
 	// Type for use as underlying render values.
-	using vector_type = double;
+	using vector_type = float;
 }
